bug: subpaths were ignored in http datasources
